% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-dors-connection.R
\name{create_context}
\alias{create_context}
\title{Create Dask-SQL/Dask/Blazing Contexts}
\usage{
create_context(CTX, SCHEDULER = FALSE)
}
\arguments{
\item{character}{choice of which context to create}
}
\description{
Create Dask-SQL/Dask/Blazing Contexts
}
