% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-dors-connection.R
\name{create_table}
\alias{create_table}
\title{Create Table}
\usage{
create_table(conn, name, df)
}
\arguments{
\item{conn}{object connection object}

\item{name}{character name of the table}

\item{df}{object data frame object or distributed data frame to write to the table}
}
\description{
Create Table
}
