% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_timeseries}
\alias{create_timeseries}
\title{Helper Function to Load Random Timeseries}
\usage{
create_timeseries(CTX, name = "timeseries")
}
\arguments{
\item{object}{DB context}

\item{character}{name of the table (default: timeseries)}
}
\description{
Helper Function to Load Random Timeseries
}
