% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-dors-connection.R
\name{DORSConnection}
\alias{DORSConnection}
\title{Create Dask SQL or BlazingSQL Connection}
\usage{
DORSConnection(CTX = "dask", SCHEDULER)
}
\arguments{
\item{CTX}{character (dask or blazing)}

\item{SCHEDULER}{character}
}
\value{
\code{DORSConnection}
}
\description{
Create Dask SQL or BlazingSQL Connection
}
