% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_dors}
\alias{install_dors}
\title{Install Python Dependencies}
\usage{
install_dors(envname = "r-reticulate")
}
\arguments{
\item{envname}{character environment name}
}
\description{
Install Python Dependencies
}
