% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-dors-connection.R
\name{dbDisconnect,DORSConnection-method}
\alias{dbDisconnect,DORSConnection-method}
\title{#' @rdname DBI
#' @inheritParams DBI::dbDisconnect}
\usage{
\S4method{dbDisconnect}{DORSConnection}(conn, ...)
}
\arguments{
\item{DORSConnection}{object DB connection object
#' @export}
}
\description{
#' @rdname DBI
#' @inheritParams DBI::dbDisconnect
}
