% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-dors-result.R
\name{run_sql}
\alias{run_sql}
\title{Run SQL through Dask/RAPIDS/Blazing Engine}
\usage{
run_sql(res)
}
\arguments{
\item{res}{object res object containing sql statement and db connection}
}
\description{
Run SQL through Dask/RAPIDS/Blazing Engine
}
